Website updates and changes. 26 & 27 December 2024 Add: (M2) (About us) Implemented. Add: (M2) (Contact us) Implemented. Add: (M2) contact/validation.php Add: (M5) donations/donations.php Add: (M5) donations/process_ddonation.php Add: (M5) account/profile/profile.php Add: (M5) account/profile/update_profile.php Add: (M4) php/submit_css_quiz.php Add: (M4) php/submit_html_quiz.php
31 March 2025
Numerous of changes that has been made in December were left unpublished.
Minor documentations changes were modified in the source codes.
As of 31 March 2025, contributors names has been changed as follows:
M1 is now M2.
M2 is now M3.
M3 is now M1.
This may be considered as the final commit for this repository as for now.
Future updates to this project may be made sometime in the foreseeable future.